# schemas/openapi.local-api.yaml
components:
  schemas:
    ArtifactMeta:
      additionalProperties: false
      properties:
        path:
          type: string
        sha256:
          type: string
        sizeBytes:
          minimum: 0
          type:
          - integer
          - 'null'
      required:
      - path
      - sha256
      type: object
    AuthContext:
      oneOf:
      - $ref: '#/components/schemas/LocalTokenAuthContext'
      - $ref: '#/components/schemas/SsoAuthContext'
    AuthContextType:
      enum:
      - local_token
      - sso
      type: string
    BaselineArtifactMeta:
      additionalProperties: false
      properties:
        path:
          $ref: '#/components/schemas/BaselineArtifactPath'
        sha256:
          type: string
        sizeBytes:
          minimum: 0
          type:
          - integer
          - 'null'
      required:
      - path
      - sha256
      type: object
    BaselineArtifactPath:
      enum:
      - veil.baseline.json
      type: string
    BaselineRequest:
      additionalProperties: false
      properties:
        outputPath:
          type:
          - string
          - 'null'
        paths:
          items:
            type: string
          type:
          - array
          - 'null'
      type: object
    BaselineResponse:
      additionalProperties: false
      properties:
        filePath:
          type: string
        findingsCount:
          minimum: 0
          type: integer
        nextAction:
          $ref: '#/components/schemas/NextAction'
        schemaVersion:
          $ref: '#/components/schemas/LocalApiSchemaVersion'
        written:
          type: boolean
      required:
      - schemaVersion
      - filePath
      - findingsCount
      - written
      - nextAction
      type: object
    BaselineStatus:
      enum:
      - none
      - new
      - suppressed
      type: string
    BindAddress:
      enum:
      - 127.0.0.1
      type: string
    BoundValue:
      oneOf:
      - format: int64
        minimum: 0
        type: integer
      - type: string
      - type: boolean
    BuildProfile:
      enum:
      - debug
      - release
      type: string
    ConfigConflict:
      additionalProperties: false
      properties:
        explanation:
          type: string
        key:
          type: string
        shadowed:
          items:
            type: string
          type: array
        winner:
          $ref: '#/components/schemas/ConfigLayerName'
      required:
      - key
      - winner
      - shadowed
      - explanation
      type: object
    ConfigLayerName:
      enum:
      - builtin
      - preset
      - org
      - repo
      - cli
      type: string
    ConfigLayerSummary:
      additionalProperties: false
      properties:
        loaded:
          type: boolean
        name:
          $ref: '#/components/schemas/ConfigLayerName'
        path:
          type:
          - string
          - 'null'
        warnings:
          items:
            type: string
          type: array
      required:
      - name
      - loaded
      - warnings
      type: object
    DoctorResponse:
      additionalProperties: false
      properties:
        bounds:
          additionalProperties:
            $ref: '#/components/schemas/BoundValue'
          propertyNames:
            type: string
          type: object
        config:
          $ref: '#/components/schemas/PolicyResponse'
        networkMode:
          $ref: '#/components/schemas/NetworkMode'
        os:
          type: string
        productVersion:
          type: string
        rulePacks:
          items:
            $ref: '#/components/schemas/DoctorRulePack'
          type: array
        rustVersion:
          type:
          - string
          - 'null'
        schemaVersion:
          $ref: '#/components/schemas/LocalApiSchemaVersion'
        warnings:
          items:
            type: string
          type: array
      required:
      - schemaVersion
      - productVersion
      - os
      - config
      - bounds
      - rulePacks
      - networkMode
      - warnings
      type: object
    DoctorRulePack:
      additionalProperties: false
      properties:
        name:
          type: string
        source:
          $ref: '#/components/schemas/RulePackSource'
        version:
          type:
          - string
          - 'null'
      required:
      - name
      - source
      type: object
    EngineMeta:
      additionalProperties: false
      properties:
        name:
          $ref: '#/components/schemas/EngineName'
        rulePacks:
          items:
            $ref: '#/components/schemas/RulePackMeta'
          type: array
        schemaVersion:
          $ref: '#/components/schemas/EngineSchemaVersion'
      required:
      - name
      - schemaVersion
      - rulePacks
      type: object
    EngineName:
      enum:
      - veil
      type: string
    EngineSchemaVersion:
      enum:
      - veil-v1
      type: string
    ErrorBody:
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        nextAction:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/NextAction'
      required:
      - code
      - message
      type: object
    ErrorCode:
      enum:
      - INVALID_REQUEST
      - UNAUTHORIZED
      - PATH_DENIED
      - NOT_FOUND
      - RUN_EXPIRED
      - RUN_TOO_LARGE
      - INTERNAL_ERROR
      type: string
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      type: object
    EvidenceArtifacts:
      additionalProperties: false
      properties:
        baseline:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BaselineArtifactMeta'
        effectiveConfig:
          $ref: '#/components/schemas/ArtifactMeta'
        reportHtml:
          $ref: '#/components/schemas/ArtifactMeta'
        reportJson:
          $ref: '#/components/schemas/ArtifactMeta'
      required:
      - reportHtml
      - reportJson
      - effectiveConfig
      type: object
    EvidenceReportSchemaVersion:
      enum:
      - veil-evidence-report-v1
      type: string
    EvidenceReportV1:
      additionalProperties: false
      properties:
        findings:
          items:
            $ref: '#/components/schemas/SafeFindingApiV1'
          type: array
        generatedAtUtc:
          type: string
        runId:
          type: string
        schemaVersion:
          $ref: '#/components/schemas/EvidenceReportSchemaVersion'
        summary:
          $ref: '#/components/schemas/EvidenceSummary'
      required:
      - schemaVersion
      - runId
      - generatedAtUtc
      - summary
      - findings
      type: object
    EvidenceSummary:
      additionalProperties: false
      properties:
        allSeverityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        coverageComplete:
          type: boolean
        effectiveFindings:
          minimum: 0
          type: integer
        severityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        suppressedFindings:
          minimum: 0
          type: integer
        suppressedSeverityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        totalFindings:
          minimum: 0
          type: integer
      required:
      - totalFindings
      - suppressedFindings
      - effectiveFindings
      - severityCounts
      - allSeverityCounts
      - suppressedSeverityCounts
      - coverageComplete
      type: object
    FindingV1:
      additionalProperties: false
      properties:
        category:
          type:
          - string
          - 'null'
        line_number:
          minimum: 1
          type: integer
        masked_snippet:
          type: string
        path:
          type: string
        rule_id:
          type: string
        score:
          format: int32
          maximum: 100
          minimum: 0
          type: integer
        severity:
          $ref: '#/components/schemas/SeverityName'
        tags:
          items:
            type: string
          type: array
      required:
      - path
      - line_number
      - rule_id
      - severity
      - score
      - masked_snippet
      - tags
      type: object
    GradeName:
      enum:
      - Low
      - Medium
      - High
      - Critical
      type: string
    LocalApiSchemaVersion:
      enum:
      - veil-pro-local-api-v1
      type: string
    LocalTokenAuthContext:
      additionalProperties: false
      properties:
        authenticated:
          type: boolean
        type:
          $ref: '#/components/schemas/AuthContextType'
      required:
      - authenticated
      - type
      type: object
    NetworkMode:
      enum:
      - local-only
      - enterprise-opt-in
      type: string
    NextAction:
      enum:
      - RESCAN
      - CHECK_TOKEN
      - NARROW_SCOPE
      - OPEN_DOCTOR
      - COMMIT_BASELINE
      - REVIEW_FINDINGS
      type: string
    PolicyResponse:
      additionalProperties: false
      properties:
        conflicts:
          items:
            $ref: '#/components/schemas/ConfigConflict'
          type: array
        effectiveRulesCount:
          minimum: 0
          type: integer
        hasOrgConfig:
          type: boolean
        layers:
          items:
            $ref: '#/components/schemas/ConfigLayerSummary'
          type: array
        orgConfigPath:
          type:
          - string
          - 'null'
        preset:
          type:
          - string
          - 'null'
        repoConfigPath:
          type:
          - string
          - 'null'
        schemaVersion:
          $ref: '#/components/schemas/LocalApiSchemaVersion'
      required:
      - schemaVersion
      - hasOrgConfig
      - effectiveRulesCount
      - layers
      - conflicts
      type: object
    PresetName:
      enum:
      - standard-jp
      - fintech-jp
      - gov-jp
      - si-vendor-jp
      - logs-jp
      type: string
    PrivacyMeta:
      additionalProperties: false
      properties:
        bind:
          $ref: '#/components/schemas/BindAddress'
        networkMode:
          $ref: '#/components/schemas/NetworkMode'
        telemetry:
          $ref: '#/components/schemas/TelemetryMode'
      required:
      - telemetry
      - networkMode
      - bind
      type: object
    ProductMeta:
      additionalProperties: false
      properties:
        buildProfile:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BuildProfile'
        commit:
          type:
          - string
          - 'null'
        name:
          $ref: '#/components/schemas/ProductName'
        version:
          type: string
      required:
      - name
      - version
      type: object
    ProductName:
      enum:
      - veil-pro
      - veil
      type: string
    ProjectSummary:
      additionalProperties: false
      properties:
        displayName:
          type: string
        hasRepoConfig:
          type: boolean
        id:
          type: string
        isCurrent:
          type: boolean
        rootPath:
          type: string
      required:
      - id
      - displayName
      - rootPath
      - isCurrent
      - hasRepoConfig
      type: object
    ProjectsResponse:
      additionalProperties: false
      properties:
        currentDir:
          type: string
        projects:
          items:
            $ref: '#/components/schemas/ProjectSummary'
          type: array
        schemaVersion:
          $ref: '#/components/schemas/LocalApiSchemaVersion'
      required:
      - schemaVersion
      - currentDir
      - projects
      type: object
    RulePackMeta:
      additionalProperties: false
      properties:
        contentSha256:
          type:
          - string
          - 'null'
        name:
          type: string
        source:
          $ref: '#/components/schemas/RulePackSource'
        version:
          type:
          - string
          - 'null'
      required:
      - name
      - source
      type: object
    RulePackSource:
      enum:
      - embedded
      - local
      - remote
      type: string
    RunMetaSchemaVersion:
      enum:
      - veil-pro-run-meta-v1
      type: string
    RunMetaV1:
      additionalProperties: false
      properties:
        artifacts:
          $ref: '#/components/schemas/EvidenceArtifacts'
        engine:
          $ref: '#/components/schemas/EngineMeta'
        extensions:
          type: object
        generatedAtUtc:
          type: string
        privacy:
          $ref: '#/components/schemas/PrivacyMeta'
        product:
          $ref: '#/components/schemas/ProductMeta'
        result:
          $ref: '#/components/schemas/RunResultMeta'
        runId:
          type: string
        schemaVersion:
          $ref: '#/components/schemas/RunMetaSchemaVersion'
      required:
      - schemaVersion
      - runId
      - generatedAtUtc
      - product
      - engine
      - result
      - artifacts
      - privacy
      type: object
    RunResultMeta:
      additionalProperties: false
      properties:
        exitCode:
          format: int32
          maximum: 2
          minimum: 0
          type: integer
        limitReached:
          type: boolean
        limitReasons:
          items:
            type: string
          type: array
        status:
          $ref: '#/components/schemas/RunStatus'
        summary:
          $ref: '#/components/schemas/EvidenceSummary'
      required:
      - status
      - exitCode
      - limitReached
      - limitReasons
      - summary
      type: object
    RunStatus:
      enum:
      - success
      - violation
      - incomplete
      - error
      type: string
    SafeFindingApiV1:
      additionalProperties: false
      properties:
        baselineFingerprint:
          type: string
        baselineStatus:
          $ref: '#/components/schemas/BaselineStatus'
        category:
          type: string
        findingId:
          type: string
        grade:
          $ref: '#/components/schemas/GradeName'
        lineNumber:
          minimum: 1
          type: integer
        maskedSnippet:
          type: string
        path:
          type: string
        ruleId:
          type: string
        score:
          format: int32
          maximum: 100
          minimum: 0
          type: integer
        severity:
          $ref: '#/components/schemas/SeverityName'
        tags:
          items:
            type: string
          type: array
      required:
      - findingId
      - baselineFingerprint
      - path
      - lineNumber
      - ruleId
      - severity
      - score
      - grade
      - maskedSnippet
      - category
      - tags
      - baselineStatus
      type: object
    ScanMode:
      enum:
      - full
      - staged
      - ci
      type: string
    ScanRequest:
      additionalProperties: false
      properties:
        baselineFile:
          type:
          - string
          - 'null'
        failOnFindings:
          minimum: 1
          type:
          - integer
          - 'null'
        failOnScore:
          format: int32
          maximum: 100
          minimum: 0
          type:
          - integer
          - 'null'
        failOnSeverity:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SeverityName'
        includeSuppressed:
          type: boolean
        mode:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ScanMode'
        paths:
          items:
            type: string
          type:
          - array
          - 'null'
        preset:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/PresetName'
      type: object
    ScanResponse:
      additionalProperties: false
      properties:
        allSeverityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        builtinSkips:
          items:
            type: string
          type: array
        coverageComplete:
          type: boolean
        effectiveFindings:
          minimum: 0
          type: integer
        expiresAtUtc:
          type: string
        findings:
          items:
            $ref: '#/components/schemas/SafeFindingApiV1'
          type: array
        limitReached:
          type: boolean
        limitReasons:
          items:
            type: string
          type: array
        runId:
          type: string
        scannedFiles:
          minimum: 0
          type: integer
        schemaVersion:
          $ref: '#/components/schemas/LocalApiSchemaVersion'
        severityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        skippedFiles:
          minimum: 0
          type: integer
        status:
          $ref: '#/components/schemas/RunStatus'
        suppressedFindings:
          minimum: 0
          type: integer
        suppressedSeverityCounts:
          $ref: '#/components/schemas/SeverityCounts'
        totalFindings:
          minimum: 0
          type: integer
      required:
      - schemaVersion
      - runId
      - status
      - scannedFiles
      - skippedFiles
      - totalFindings
      - suppressedFindings
      - effectiveFindings
      - coverageComplete
      - severityCounts
      - allSeverityCounts
      - suppressedSeverityCounts
      - limitReached
      - limitReasons
      - builtinSkips
      - findings
      - expiresAtUtc
      type: object
    SeverityCounts:
      additionalProperties: false
      properties:
        Critical:
          minimum: 0
          type: integer
        High:
          minimum: 0
          type: integer
        Low:
          minimum: 0
          type: integer
        Medium:
          minimum: 0
          type: integer
      required:
      - Low
      - Medium
      - High
      - Critical
      type: object
    SeverityName:
      enum:
      - Low
      - Medium
      - High
      - Critical
      type: string
    SsoAuthContext:
      additionalProperties: false
      properties:
        authenticated:
          type: boolean
        email:
          type: string
        enterpriseOptIn:
          type: boolean
        name:
          type:
          - string
          - 'null'
        type:
          $ref: '#/components/schemas/AuthContextType'
      required:
      - authenticated
      - type
      - email
      - enterpriseOptIn
      type: object
    TelemetryMode:
      enum:
      - none
      type: string
  securitySchemes:
    bearerAuth:
      bearerFormat: opaque
      scheme: bearer
      type: http
info:
  description: Generated from Rust DTOs. Do not hand-edit in implementation.
  title: Veil Pro Local API
  version: 1.0.0-local
openapi: 3.0.3
paths:
  /api/baseline:
    post:
      operationId: openapi_write_baseline
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaselineRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineResponse'
          description: Baseline result
      security:
      - bearerAuth: []
      tags: []
  /api/doctor:
    get:
      operationId: openapi_get_doctor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
          description: Doctor diagnostics
      security:
      - bearerAuth: []
      tags: []
  /api/me:
    get:
      operationId: openapi_get_me
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthContext'
          description: Auth context
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
      security:
      - bearerAuth: []
      tags: []
  /api/policy:
    get:
      operationId: openapi_get_policy
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
          description: Effective policy
      security:
      - bearerAuth: []
      tags: []
  /api/projects:
    get:
      operationId: openapi_list_projects
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsResponse'
          description: Project list
      security:
      - bearerAuth: []
      tags: []
  /api/runs/{runId}:
    get:
      operationId: openapi_get_run_meta
      parameters:
      - description: Run identifier
        in: path
        name: runId
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunMetaV1'
          description: Run metadata
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Run expired
      security:
      - bearerAuth: []
      tags: []
  /api/runs/{runId}/evidence.zip:
    get:
      operationId: openapi_export_evidence
      parameters:
      - description: Run identifier
        in: path
        name: runId
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/zip: {}
          description: Evidence ZIP
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Run expired
      security:
      - bearerAuth: []
      tags: []
  /api/scan:
    post:
      operationId: openapi_scan_project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
          description: Scan response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Invalid request
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Path denied
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Run too large
      security:
      - bearerAuth: []
      tags: []
servers:
- url: http://127.0.0.1:{port}
  variables:
    port:
      default: '3000'